Make search work on other platforms
